Improve role sync performance. Fixes #306
Before this comment, as soon as any role was changed, the permissions for
all student-quizzes were changed.

After this, when the roles change, then we note that time in a cache.

Then, when a user tries to access a StudentQuiz, if the roles have
changed since we last synced permissions for this StudentQuiz,
then we re-sync them (and note the fact in the cache).

Because of the way the cache is used, if the caches are purged, then
we will re-sync the permissions for each StudentQuiz the next time a
user tries to access it.