Bad data can break the 4.0 upgrade
In our database, we seem to have some questions where the corresponding
question_cagtegory does not exist (presumably due to an old bug).

That should never happen, but clearly it does, so we need to handle it.
With this change, it will leave some data in the database with
studentquizid = 0 or studentquizquestionid = 0, but that bad data was
there already.